change embedding type from float array to number list, and double in vector repo

+++ src/main/java/uk/ac/ebi/protvar/dto/VectorSearchResult.java
@@ -1,3 +1,3 @@
 package uk.ac.ebi.protvar.dto;
 
-public record VectorSearchResult(String accession, String sourceType, String sourceText, Float distance) {}
+public record VectorSearchResult(String accession, String sourceType, String sourceText, Double distance) {}
